Fix connect-domain placement & community controller.
If connect-domain is in front of the http middleware (especially the
body parser) the whole application doesn't work.
It's not clear to me why exactly but I'll figure this out.
It might be something with the chunk parsing in the express/connect
bodyParser (req.on('end') is never reached).

Community controller translation call has been fixed.